sqlite index parse

+++ src/core/field/migration.rs
@@ -7,5 +7,5 @@ pub(crate) struct FieldMigration {
     pub(crate) version: Option<String>,
     pub(crate) default: Option<Value>,
     pub(crate) action: Option<Pipeline>,
-    pub(crate) priority: Option<usize>,
+    pub(crate) priority: Option<i64>,
 }
